RD-1598 - schema renamed to json file
